Move match creation to MatchService
